challenge-4/agent-workflow/app/pyproject.toml
[project]
name = "FastAPI-starter"
version = "0.1.0"
description = "FastAPI starter"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "agent-framework>=0.1.0",
    "agent-framework-azure-ai>=0.1.0",
    "azure-identity>=1.15.0",
    "azure-cosmos>=4.5.0",
    "python-dotenv>=1.0.0",
    "opentelemetry-distro>=0.59b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.38.0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "agent-framework-a2a>=1.0.0b260107",
]

[tool.uv]
prerelease = "allow"
